+++ src/main/java/com/wuwu/worldofzuulwuwu/service/PlayerService.java
@@ -1,28 +1,9 @@
 package com.wuwu.worldofzuulwuwu.service;
 
 import com.wuwu.worldofzuulwuwu.entity.Player;
-import com.wuwu.worldofzuulwuwu.mapper.PlayerDao;
 
-public class PlayerService {
-
-  private PlayerDao playerDao;
-
-  public Boolean register(Player player){
-    Player player1 = playerDao.save(player);
-    if(player1!=null){
-      return true;
-    }else{
-      return false;
-    }
-  }
-
-  public Boolean login(Player player){
-    Player player1 = playerDao.findByName(player.getName());
-    if(player1.getPassword().equals(player.getPassword())){
-      return true;
-    }else{
-      return false;
-    }
-  }
+public interface PlayerService {
+  Boolean register(Player player);
+  Boolean login(Player player);
 
 }
